% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulates a single trial}
\usage{
sim_trial(
  design,
  tox_rates,
  cohortsize = 1,
  start = 1,
  parallel_recruit = TRUE,
  futime,
  accrual
)
}
\arguments{
\item{design}{The dose-finding design which is employed. Possible options are "RollingSix" for Rolling Six.}

\item{tox_rates}{Vector of true toxicity rates for each dose level in ascending order}

\item{cohortsize}{Number of patients in a cohort}

\item{start}{Dose level to be used at the beginning of the trial}

\item{parallel_recruit}{Whether recruitment is parallel or sequential. If FALSE recruitment is sequential i.e. recruitment is suspended until all DLT results are known}

\item{futime}{Follow-up time in days}

\item{accrual}{Expected number of patients to be recruited in a year}
}
\value{
res_dlt Data frame with dose level and DLT result for each patient

res_time Data frame with important time points in days since trial start for each patient
}
\description{
Simulates a single trial
}
