% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_trials}
\alias{sim_trials}
\title{Simulates multiple trials}
\usage{
sim_trials(
  design,
  tox_rates,
  cohortsize = 1,
  start = 1,
  parallel_recruit = TRUE,
  futime,
  accrual,
  nsim,
  seed = 1
)
}
\arguments{
\item{design}{Thedose-finding design which is employed. Possible options are "RollingSix" for Rolling Six.}

\item{tox_rates}{Vector of true toxicity rates for each dose level in ascending order}

\item{cohortsize}{Number of patients in a cohort}

\item{start}{Dose level to be used at the beginning of the trial}

\item{parallel_recruit}{Whether recruitment is parallel or sequential. If FALSE recruitment is sequential i.e. recruitment is suspended until all DLT results are known}

\item{futime}{Follow-up time in days}

\item{accrual}{Expected number of patients to be recruited in a year}

\item{nsim}{Number of simulated trials}

\item{seed}{Seed for simulation}
}
\value{
res Data frame with key numbers for each trial
}
\description{
Simulates multiple trials
}
