% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{summary_trial}
\alias{summary_trial}
\title{Summarizes the result of a single simulated trial}
\usage{
summary_trial(res)
}
\arguments{
\item{res}{Output from \code{sim_trial}}
}
\description{
Summarizes the result of a single simulated trial
}
